+++ lib/src/utils/file_saver.dart
@@ -1,0 +1,2 @@
+export 'file_saver_stub.dart'
+    if (dart.library.io) 'file_saver_io.dart';